2020.07.27 // 김재성 // 종합
package 구성
=====controller=====
HomeController.java
====================

=======model========
Coordinate.java
stationXY.java
====================

=======service======
(interface)Mapservice.java
MapServiceImpl.java
====================

========json========
JsonPsraing.java
====================

// 수정 사항
1. addresTest 프로젝트와 meetnow 프로젝트 통합

2.MapServiceImpl.getStationCoord(...)
	keyword api 호출 ==> category api 호출.
	page, size, sort 속성 추가.

3.뒤로가기 세션 처리(장소명, 주소, 좌표값)
	문제점 :뒤로 가기했을 때
		리스트가 바로 출력되지 않고
		새로고침 필요함.

+++ addrestTest/src/main/java/com/mycompany/myapp/service/MapServiceImpl.java
@@ -92,15 +92,15 @@ public class MapServiceImpl implements MapService {
 	// 출발지 좌표 값의 평균.
 	public Coordinate getCenter(String[] x, String[] y) {
 		Coordinate coor = new Coordinate();
-		float n = (float)x.length;
-		float sumX =0;
-		float sumY =0;
+		double n = (double)x.length;
+		double sumX =0;
+		double sumY =0;
 		for(int i=0; i<x.length; i++) {
-			sumX+=Float.parseFloat(x[i]);
-			sumY+=Float.parseFloat(y[i]);
+			sumX+=Double.parseDouble(x[i]);
+			sumY+=Double.parseDouble(y[i]);
 		}
-		coor.setX(Float.toString(sumX/n));
-		coor.setY(Float.toString(sumY/n));
+		coor.setX(Double.toString(sumX/n));
+		coor.setY(Double.toString(sumY/n));
 		return coor;
 	}
 }
